% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{bibliography_entries}
\alias{bibliography_entries}
\title{Print bibliography section}
\usage{
bibliography_entries(file, startlabel = NULL, endlabel = NULL)
}
\arguments{
\item{file}{A path to a bibliography file understood by \code{\link[rmarkdown:pandoc_citeproc_convert]{rmarkdown::pandoc_citeproc_convert()}}.}

\item{startlabel}{Defunct.}

\item{endlabel}{Defunct.}
}
\value{
A dataset representing the bibliographic entries, suitable for
generating a reference section in a document.
}
\description{
Given a bibliography file, this function will generate bibliographic entries
for one or more types of bib entry.
}
\author{
Mitchell O'Hara-Wild & Rob J Hyndman
}
