% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csl_types.R
\name{csl_name}
\alias{csl_name}
\title{A list of names conforming to the CSL schema}
\usage{
csl_name(x)
}
\arguments{
\item{x}{A list of names conforming to the structure outlined in the CSL schema}
}
\description{
This class provides helper utilities to display, sort, and select attributes
from a name in the CSL format.
}
\seealso{
\url{https://citeproc-js.readthedocs.io/en/latest/csl-json/markup.html#name-fields}
}
