% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{cv}
\alias{cv}
\title{Output formats for CVs}
\usage{
cv(template = "hyndman", ...)
}
\arguments{
\item{template}{Name of template to be used.}

\item{\dots}{Arguments passed to \code{\link[bookdown]{pdf_document2}}.}
}
\value{
An R Markdown output format object.
}
\description{
Each function is a wrapper for \code{\link[bookdown]{pdf_document2}} to
produce curriculum vitae
}
\author{
Rob J Hyndman
}
