% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_document.R
\name{cv_document}
\alias{cv_document}
\title{Output format for vitae}
\usage{
cv_document(...)
}
\arguments{
\item{...}{Arguments passed to bookdown::pdf_document2()}
}
\description{
This output format provides support for including LaTeX dependencies and
bibliography entries in extension of the bookdown::pdf_document2() format.
}
