% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{cv_formats}
\alias{cv_formats}
\alias{hyndman}
\alias{twentyseconds}
\alias{awesomecv}
\alias{moderncv}
\title{Output formats for CVs}
\usage{
hyndman(...)

twentyseconds(...)

awesomecv(...)

moderncv(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[vitae]{cv_document}}.}
}
\value{
An R Markdown output format object.
}
\description{
Each function is a wrapper for \code{\link[vitae]{cv_document}} to
produce curriculum vitae
}
\author{
Mitchell O'Hara-Wild and Rob J Hyndman
}
