% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honors.R
\name{honor_entries}
\alias{honor_entries}
\title{Create honor entries}
\usage{
honor_entries(data, what, when = NA, with = NA, where = NA, .protect = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} containing the honor entries}

\item{what}{Honor title/achievement}

\item{when}{Date or time period of honor}

\item{with}{Organization/institution providing the honor}

\item{where}{Optional location of the honor}

\item{.protect}{Logical, if TRUE (default) protects special LaTeX characters in input}
}
\value{
Object of class \code{vitae_honor}
}
\description{
Create honor entries
}
