% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{hyndman}
\alias{hyndman}
\alias{twentyseconds}
\title{Output formats for CVs}
\usage{
hyndman(...)

twentyseconds(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[bookdown]{pdf_document2}}.}
}
\value{
An R Markdown output format object.
}
\description{
Each function is a wrapper for \code{\link[bookdown]{pdf_document2}} to
produce curriculum vitae
}
\author{
Rob J Hyndman
}
