% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{insert_preview}
\alias{insert_preview}
\title{Include a preview of the CV template output for documentation}
\usage{
insert_preview(template)
}
\arguments{
\item{template}{Name of the template}
}
\description{
Include a preview of the CV template output for documentation
}
\keyword{internal}
