% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleton.R
\name{middleton}
\alias{middleton}
\title{Middleton CV template}
\usage{
middleton(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[vitae]{cv_document}}.}
}
\value{
An R Markdown output format object.
}
\description{
Produces a CV using the style used in Rob Hyndman's CV:
https://robjhyndman.com/hyndsight/cv/
}
\author{
Rob J Hyndman & Mitchell O'Hara-Wild with edits by KMM
}
