% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onecoltable.R
\name{onecoltable_entries}
\alias{onecoltable_entries}
\title{Create onecoltable entries}
\usage{
onecoltable_entries(data, details, .protect = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} containing the onecoltable text entries}

\item{details}{the column name in \code{data} that contains the details
for each entry}

\item{.protect}{Logical, if TRUE (default) protects special LaTeX characters in input}
}
\value{
Object of class \code{vitae_onecoltable}
}
\description{
Create onecoltable entries
}
