% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plain.R
\name{plain_entries}
\alias{plain_entries}
\title{Create plain entries}
\usage{
plain_entries(data, details, .protect = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} containing the plain text entries}

\item{details}{the column name in \code{data} that contains the details
for each entry}

\item{.protect}{Logical, if TRUE (default) protects special LaTeX characters in input}
}
\value{
Object of class \code{vitae_plain}
}
\description{
Create plain entries
}
