% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{printbibliography}
\alias{printbibliography}
\title{Print bibliography section}
\usage{
printbibliography(bib, category = c("Article"),
  title = "Refereed journal papers", sorting = "ynt",
  startlabel = NULL, endlabel = NULL)
}
\arguments{
\item{bib}{Name of bib file}

\item{category}{A character vector specifying the types of bib entries to include}

\item{title}{Character string giving section title}

\item{sorting}{Character stringing specifying how entries should be sorted. Default is "ynt" meaning
sort by year first, then name, then title.}

\item{startlabel}{Optional label for first reference in the section}

\item{endlabel}{Optional label for last reference in the section}
}
\value{
Prints bibliographic entries
}
\description{
Given a bib file, this function will generate bibliographic entries for one or more types of bib entry.
}
\author{
Rob J Hyndman
}
