% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{mutate}
\alias{transmute}
\alias{group_by}
\alias{summarise}
\alias{rename}
\alias{arrange}
\alias{select}
\alias{filter}
\alias{distinct}
\alias{slice}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:\%>\%]{\%>\%}}, \code{\link[dplyr:arrange]{arrange()}}, \code{\link[dplyr:distinct]{distinct()}}, \code{\link[dplyr:filter]{filter()}}, \code{\link[dplyr:group_by]{group_by()}}, \code{\link[dplyr:mutate]{mutate()}}, \code{\link[dplyr:rename]{rename()}}, \code{\link[dplyr:select]{select()}}, \code{\link[dplyr:slice]{slice()}}, \code{\link[dplyr:summarise]{summarise()}}, \code{\link[dplyr:transmute]{transmute()}}}
}}

