% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_monocle3.R
\name{run_fun}
\alias{run_fun}
\title{Run Monocle 3}
\usage{
run_fun(expression, parameters, priors, verbose, seed)
}
\arguments{
\item{expression}{Expression matrix}

\item{parameters}{Parameters}

\item{priors}{Priors}

\item{verbose}{Verbosity level}

\item{seed}{Random seed}
}
\description{
Run Monocle 3
}
