% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills.R
\name{skill_entries}
\alias{skill_entries}
\title{Create skill entries}
\usage{
skill_entries(data, type, details, .protect = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} containing the skill entries}

\item{type}{Skill type/category}

\item{details}{Skill details/descriptions}

\item{.protect}{Logical, if TRUE (default) protects special LaTeX characters in input}
}
\value{
Object of class \code{vitae_skill}
}
\description{
Create skill entries
}
