% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_monocle3.R
\name{ti_monocle3}
\alias{ti_monocle3}
\title{Infer a trajectory using Monocle 3}
\usage{
ti_monocle3(
  num_dim = 50L,
  max_components = 2L,
  reduction_method = "UMAP",
  cluster_method = "leiden",
  k = 20L
)
}
\arguments{
\item{num_dim}{The dimensionality of the preprocessing reduced space. Domain: U(10, 100). Default: 50. Format: integer.}

\item{max_components}{The dimensionality of the reduced space. Domain: U(2, 20). Default: 2. Format: integer.}

\item{reduction_method}{A character string specifying the algorithm to use for dimensionality reduction.Currently "UMAP", "tSNE", "PCA" and "LSI" are supported. Domain: {UMAP, tSNE, PCA, LSI}. Default: UMAP. Format: character.}

\item{cluster_method}{Indicates the clustering method to use. Domain: {leiden, louvain}. Default: leiden. Format: character.}

\item{k}{Number of nearest neighbors to use when creating the k nearest neighbor graph for Louvain/Leiden clustering.k is related to the resolution of the clustering result, a bigger k will result in lower resolutionand vice versa. Domain: U(5, 50). Default: 20. Format: integer.}
}
\description{
Infer a trajectory using Monocle 3
}
\examples{
dataset <- data <- dyntoy::generate_dataset(
  num_cells = 99,
  num_features = 101,
  model = "tree",
  normalise = FALSE
)
model <- dynwrap::infer_trajectory(dataset, ti_monocle3())

}
