% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twocoltable.R
\name{twocoltable_entries}
\alias{twocoltable_entries}
\title{Create twocoltable entries}
\usage{
twocoltable_entries(data, details, .protect = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} containing the twocoltable text entries}

\item{details}{the column name in \code{data} that contains the details
for each entry}

\item{.protect}{Logical, if TRUE (default) protects special LaTeX characters in input}
}
\value{
Object of class \code{vitae_twocoltable}
}
\description{
Create twocoltable entries
}
