% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{write_packages_bib}
\alias{write_packages_bib}
\title{Create bib file for R packages}
\usage{
write_packages_bib(pkglist, file)
}
\arguments{
\item{pkglist}{A character vector containing package names.}

\item{file}{Name of bib file to write to.}
}
\value{
Nothing.
}
\description{
Given a list of R packages, this function will generate a bib file using CRAN information if
it is available. Otherwise, it uses github. Packages that are not on CRAN or github will generate
errors. Also packages that have not been installed using either `install.packages()` or `devtools::install_github()`
will cause errors.
}
\author{
Rob J Hyndman
}
