% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csl_name.R
\name{csl_name}
\alias{csl_name}
\title{A name variable conforming to the CSL schema}
\usage{
csl_name(
  family = NULL,
  given = NULL,
  dropping_particle = NULL,
  non_dropping_particle = NULL,
  suffix = NULL,
  comma_suffix = NULL,
  static_ordering = NULL,
  literal = NULL,
  parse_names = NULL
)
}
\arguments{
\item{family}{The family name}

\item{given}{The given name}

\item{dropping_particle, non_dropping_particle, comma_suffix, static_ordering, literal, parse_names}{Additional
name variable properties as described in the schema.}
}
\description{
This class provides helper utilities to display, sort, and select attributes
from a name in the CSL format.
}
\seealso{
\url{https://citeproc-js.readthedocs.io/en/latest/csl-json/markup.html#name-fields}
}
