% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_document.R
\name{cv_document}
\alias{cv_document}
\title{Output format for vitae}
\usage{
cv_document(
  ...,
  pandoc_args = NULL,
  pandoc_vars = NULL,
  base_format = rmarkdown::pdf_document
)
}
\arguments{
\item{...}{Arguments passed to rmarkdown::pdf_document().}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{pandoc_vars}{Pandoc variables to be passed to the template.}
}
\description{
This output format provides support for including LaTeX dependencies and
bibliography entries in extension of the \code{rmarkdown::pdf_document()} format.
}
