% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awesomecv.R, R/formats.R, R/hyndman.R,
%   R/latexcv.R, R/moderncv.R, R/twentyseconds.R
\name{awesomecv}
\alias{awesomecv}
\alias{cv_formats}
\alias{hyndman}
\alias{latexcv}
\alias{moderncv}
\alias{twentyseconds}
\title{Output formats for CVs}
\usage{
awesomecv(..., latex_engine = "xelatex")

hyndman(...)

latexcv(..., theme = c("classic", "modern", "rows", "sidebar", "two_column"))

moderncv(
  ...,
  theme = c("casual", "classic", "oldstyle", "banking", "fancy"),
  latex_engine = "pdflatex"
)

twentyseconds(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[vitae]{cv_document}}.}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", and "xelatex".}

\item{theme}{The theme used for the template.}
}
\value{
An R Markdown output format object.
}
\description{
Each function is a wrapper for \code{\link[vitae]{cv_document}} to
produce curriculum vitae
}
\author{
Mitchell O'Hara-Wild and Rob J Hyndman
}
