% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitae-package.R
\docType{package}
\name{vitae-package}
\alias{vitae}
\alias{vitae-package}
\title{vitae: Curriculum Vitae for R Markdown}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Provides templates and functions to simplify the production and maintenance of curriculum vitae.
}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

}
\keyword{internal}
