% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitae-package.R
\docType{package}
\name{vitae-package}
\alias{vitae}
\alias{vitae-package}
\title{vitae: Curriculum Vitae for R Markdown}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Provides templates and functions to simplify the production and maintenance of curriculum vitae.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/vitae/}
  \item Report bugs at \url{https://github.com/ropenscilabs/vitae/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (0000-0001-6729-7695)

Authors:
\itemize{
  \item Rob Hyndman \email{Rob.Hyndman@monash.edu} (0000-0002-2140-5352)
}

}
\keyword{internal}
