% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitae-package.R
\docType{package}
\name{vitae-package}
\alias{vitae}
\alias{vitae-package}
\title{vitae: Curriculum Vitae for R Markdown}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides templates and functions to simplify the production and maintenance of curriculum vitae.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.mitchelloharawild.com/vitae/}
  \item \url{https://github.com/ropenscilabs/vitae}
  \item Report bugs at \url{https://github.com/ropenscilabs/vitae/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

Authors:
\itemize{
  \item Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})
}

Other contributors:
\itemize{
  \item Yihui Xie (\href{https://orcid.org/0000-0003-0645-5666}{ORCID}) [contributor]
}

}
\keyword{internal}
